FDP-697 ~ Changes MQTT message payload from String to byte array
MQTT message payloads are determined by the use case. Not in all
scenarios do the bytes in the payload match the UTF-8 bytes of a String.
They may also be bytes for a zipped representation. To get rid of issues
by unexpected payloads, these changes consider the payload only as byte
array, without interpretation.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>